set up filesystem
