+++ client/src/components/Book.jsx
@@ -1,0 +1,1 @@
+// individual book view
