+++ client/src/components/SearchResults.jsx
@@ -1,0 +1,1 @@
+// children: postForm
