+++ client/src/helpers/hooks/useApplicationData.jsx
@@ -1,0 +1,1 @@
+// in-app states
