+++ client/src/routes/Profile.jsx
@@ -1,0 +1,1 @@
+// profile details, children: my posts, book list x4(to read, reading, read, favorites), postform
